test(range): migrate legacy events
